% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emforreR.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of R functions (\code{emforreR}) and Rcpp functions (\code{emforreC}).
}
\examples{
\dontrun{
data(reaction_time)
attach(reaction_time)
tm <- microbenchmark::microbenchmark(
  estR = emforreR(reaction_time),
  estC = emforreC(reaction_time),
  times = 5
)
print(summary(tm)[,c(1,3,5,6)])
}
}
