% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emforreC}
\alias{emforreC}
\title{EM Algorithm for Repeated Experiment (Rcpp Version)}
\usage{
emforreC(data)
}
\arguments{
\item{data}{the data frame to be analyzed}
}
\value{
a list containing the estimation results, iteration times and optimal log likelihood.
}
\description{
Inference on repeated experiment using EM algorithm (Rcpp version).
}
\examples{
\dontrun{
data(reaction_time)
attach(reaction_time)
result <- emforreC(reaction_time)
alpha.hat <- result$alpha; lambda.hat <- result$lambda
beta.hat <- result$beta; mu.hat <- result$mu; gamma.hat <- result$gamma
sigma.hat <- result$sigma; tau.hat <- result$tau
}
}
