% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emforreR.R
\name{emforreR}
\alias{emforreR}
\title{EM Algorithm for Repeated Experiment (R version)}
\usage{
emforreR(data)
}
\arguments{
\item{data}{the data frame to be analyzed}
}
\value{
a list containing the estimation results, iteration times and optimal log likelihood.
}
\description{
Inference on repeated experiment using EM algorithm (R version).
}
\examples{
\dontrun{
data(reaction_time)
attach(reaction_time)
result <- emforreR(reaction_time)
alpha.hat <- result$alpha; lambda.hat <- result$lambda
beta.hat <- result$beta; mu.hat <- result$mu; gamma.hat <- result$gamma
sigma.hat <- result$sigma; tau.hat <- result$tau
}
}
