% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indregclu.R
\name{indregclu}
\alias{indregclu}
\title{Individualized Regression For Clustering Subjects}
\usage{
indregclu(data, N, K)
}
\arguments{
\item{data}{the data frame to be analyzed}

\item{N}{the number of subjects}

\item{K}{the number of clustering centers}
}
\value{
a list containing the clustering result and the coefficients for individualized regression
}
\description{
A clustering method based on individualized regression
}
\examples{
\dontrun{
data(sim_data)
attach(sim_data)
result <- indregclu(sim_data, N = 200, K = 2)
clu_res <- result$clus_id
reg_res <- result$coef
}
}
