% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emforreR.R
\name{reaction_time}
\alias{reaction_time}
\title{A real dataset}
\description{
A real dataset from medical test.
}
\examples{
\dontrun{
data(reaction_time)
attach(reaction_time)
result <- emforreR(reaction_time)
}
}
