% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indregclu.R
\name{sim_data}
\alias{sim_data}
\title{A simulated dataset}
\description{
A simulated dataset used to test the individualized regression for clustering subjects
}
\examples{
\dontrun{
data(sim_data)
attach(sim_data)
result <- indregclu(sim_data, N = 200, K = 2)
}
}
