\name{StatComp22034-package}
\alias{StatComp22034-package}
\alias{StatComp22034}
\docType{package}
\title{
  This package provides several functions in 'Statistical Computing' course and some others.
}
\description{
  All 11 homework of mine are included in the package, which are from the book "Statistical Computing with R" and "Advanced R". Besides, two functions concerning about regression with penalty and gaussian mixture model are included.
}
\author{
Hu Weitao

Maintainer: Hu Weitao <watthu@mail.ustc.edu.cn>
}
\references{
  Belin, T. R., & Rubin, D. B. (1995). The analysis of repeated‐measures data on schizophrenic reaction times using mixture models. Statistics in medicine, 14(8), 747-768.
  
  Gelman, A., Carlin, J. B., Stern, H. S., and Rubin, D. B. (1995). Bayesian data analysis. Chapman and Hall/CRC.
  
  Rizzo, M. L. (2019). Statistical computing with R. Chapman and Hall/CRC.
  
  Wickham, H. (2019). Advanced r. CRC press.
  
  Huang, J., Jiao, Y., Liu, Y., & Lu, X. (2018). A constructive approach to l0 penalized regression. The Journal of Machine Learning Research, 19(1), 403-439.
  
  Tang, X., Xue, F., & Qu, A. (2021). Individualized multidirectional variable selection. Journal of the American Statistical Association, 116(535), 1280-1296.
}
\keyword{ package }
\seealso{
  [my_lecture_notes](https://bookdown.org/watthu16/_Book_StatCompWithR/)
}
